[+] Add generated for KC705 axi clock converter to provide 200 MHz access from the 40 MHz system bus. Update example to monitor and print DDR init done flag.

+++ examples/bootrom_tests/src/test_l2coherence.c
@@ -28,7 +28,7 @@ void test_l2coherence(void) {
     uint32_t cohtestcnt = 0;
     uint32_t mod40;
 
-    pnp->fwdbg1 = &cohtestcnt;     // shared variable in local stack
+    pnp->fwdbg1 = (uint64_t)&cohtestcnt;     // shared variable in local stack
     pnp->fwdbg2 = fw_get_cpuid();  // to debug in RTL and see CPU index
 
     printf_uart("%s", "L2.Coherence . .");
